// frontend-web/types/scout.ts
export interface Scout {
  id: string;
  numero: number;
  nombre: string;
  dni?: number;
  fecha_nacimiento: string;
  tel_emergencias?: string;
  tel_emergencias_2?: string;
  tel_propio?: string;
  religion?: string;
  categoria_miembro?: string;
  dia_cumple?: number;
  edad: number;
  va: boolean;
  pdf_url?: string;
  equipo: string;
  promesa?: string;
  etapa?: string;
  elemento?: "Fuego" | "Tierra" | "Agua" | "Aire";
  raid?: string;
  cordi?: string;
  guardian?: string;
  diario_marcha?: string;
  rol_comunidad?: string;
  tada_puede: boolean;
  tada_tiene: boolean;
  tada_estado: "Tiene" | "Puede hacer" | "No puede";
  mistica?: string;
  ingreso_movimiento?: "Manada" | "Unidad" | "Caminantes";
  distrito?: boolean;
  afiliacion?: boolean;
  cuota_social?: Record<string, boolean | null>;
  retiro_solo?: string;
  ddjj_salud?: string;
  uso_imagen?: string;
  otros_documentos?: Array<{ name: string; url: string }>;
  logo_elemento?: string;
  logo_promesa?: string;
  is_active: boolean;
}